This isn't working...
The input to the scanner is just a list of files form a glob,
and has no sense of order. Files that are prefixed are parsed
before we hit the index file doing the forwarding.

Maybe a path forward (heh) is to adjust the input to the scanner
to never include partials?